LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-subclasses/cssRotate.tentative-expected.txt

CONSOLE MESSAGE: TypeError: CSS.number is not a function. (In 'CSS.number(10)', 'CSS.number' is undefined)

Harness Error (FAIL), message = TypeError: CSS.number is not a function. (In 'CSS.number(10)', 'CSS.number' is undefined)
